+++ .storybook/main.js
@@ -1,3 +1,5 @@
+const path = require('path')
+
 module.exports = {
   stories: ['../stories/**/*.stories.(tsx|mdx)'],
   addons: [
@@ -7,4 +9,33 @@ module.exports = {
     '@storybook/addon-knobs',
     '@storybook/addon-docs',
   ],
+  webpackFinal: (config) => {
+
+    // ===================
+    /**
+     * modify storybook default config
+     * remove svg default file-loader
+     * use both @svgr/webpack and file-loader
+     */
+    const fileLoaderRule = config.module.rules.find(rule => {
+      try {
+        if (rule.test.test('.svg')) {
+          return true
+        }
+      } catch (error) {
+      }
+      return false
+    });
+    fileLoaderRule.test = /\.(ico|jpg|jpeg|png|gif|eot|otf|webp|ttf|woff|woff2|cur|ani|pdf)(\?.*)?$/
+    config.module.rules.push({
+      test: /\.svg$/,
+      use: ['@svgr/webpack', {
+        loader: 'file-loader',
+        options: { name: 'static/media/[name].[hash:8].[ext]', esModule: false },
+      }],
+    })
+    // ===================
+
+    return config
+  }
 }