+++ .storybook/webpack.config.js
@@ -15,7 +15,7 @@ module.exports = ({ config }) => {
         options: {
           // Provide the path to your tsconfig.json so that your stories can
           // display types from outside each individual story.
-          tsconfigPath: path.resolve(__dirname, "../tsconfig.json"),
+          tsconfigPath: path.resolve(__dirname, "./tsconfig.json"),
         },
       },
     ],